style: reorganize in package
