+++ src/main/kotlin/provider/query/BaseRepoSearchQuery.kt
@@ -1,5 +1,6 @@
-package provider
+package provider.query
 
+import provider.repository.Repository
 import java.net.URL
 
 private const val EXPECTED_PATH_ARGS = 2
