+++ src/main/kotlin/provider/repository/Repository.kt
@@ -1,4 +1,4 @@
-package provider
+package provider.repository
 
 import java.io.File
