+++ src/test/kotlin/provider/RepositoryTest.kt
@@ -3,6 +3,7 @@ package provider
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.ints.shouldBeExactly
+import provider.repository.clonedReposDirectoryPath
 import java.io.File
 import java.net.URL
